Move simulation stack helper into its own module
- Extract `getSimulationStack` from `fetchSimulationStack.ts`
- Keep simulation stack version handling and address lookup isolated
